refactor: move session to shared types pkg
